automatically backed up by learn

+++ app.rb
@@ -13,7 +13,7 @@ class App < Sinatra::Base
   end
 
   get 'say/:number/:phrase' do
-    num = params[:number].to_f
+    num = params[:number].to_i
     num.times {
       puts params[:phrase]
     }
